LibraryStrollRFW/tmp/LibraryDRF/LibraryApp/views.py
from rest_framework import generics
from django.shortcuts import render
from .models import *
from .serializers import BookSerializer

class BookAPIView(generics.ListAPIView):
    queryset = Book.objects.all()
    serializer_class = BookSerializer
# Create your views here.
